Fix bugs from SalaryDto, Department and Role

+++ src/main/java/com/ercanbeyen/employeemanagementsystem/service/impl/DepartmentServiceImpl.java
@@ -5,7 +5,8 @@ import com.ercanbeyen.employeemanagementsystem.dto.DepartmentDto;
 import com.ercanbeyen.employeemanagementsystem.dto.request.UpdateDepartmentRequest;
 import com.ercanbeyen.employeemanagementsystem.entity.Department;
 
-import com.ercanbeyen.employeemanagementsystem.exception.DepartmentNotFound;
+import com.ercanbeyen.employeemanagementsystem.exception.DataConflict;
+import com.ercanbeyen.employeemanagementsystem.exception.DataNotFound;
 import com.ercanbeyen.employeemanagementsystem.repository.DepartmentRepository;
 import com.ercanbeyen.employeemanagementsystem.service.DepartmentService;
 
@@ -38,7 +39,7 @@ public class DepartmentServiceImpl implements DepartmentService {
         String departmentName = updateDepartmentRequest.getName();
 
         Department departmentInDb = departmentRepository.findById(id).orElseThrow(
-                () -> new DepartmentNotFound("Department called " + departmentName + " is not found")
+                () -> new DataNotFound("Department called " + departmentName + " is not found")
 
         );
 
@@ -52,12 +53,25 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     public Department assignDepartment(String departmentName) {
         return departmentRepository.findByName(departmentName).orElseThrow(
-                () -> new DepartmentNotFound("Department called " + departmentName + " is not found")
+                () -> new DataNotFound("Department called " + departmentName + " is not found")
         );
     }
 
     @Override
     public void deleteDepartment(int id) {
+        Department department = departmentRepository.findById(id).orElseThrow(
+                () -> new DataNotFound("Department with id " + id + " is not found")
+        );
+
+        int numberOfEmployee = department.getEmployees().size();
+
+        if (numberOfEmployee > 0) {
+            throw new DataConflict(
+                    "Department called " + department.getName() + " could not be deleted, because it contains "
+                            + numberOfEmployee + " employee(s)"
+            );
+        }
+
         departmentRepository.deleteById(id);
     }
 
@@ -69,7 +83,7 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     public DepartmentDto getDepartment(int id) {
         Department department = departmentRepository.findById(id).orElseThrow(
-                () -> new DepartmentNotFound("Department with id " + id + " is not found")
+                () -> new DataNotFound("Department with id " + id + " is not found")
         );
 
         return convertDepartmentToDepartmentDto(department);
